+++ src/Vista/PanelPrimerUsoCrearAdmin.java
@@ -181,8 +181,8 @@ public class PanelPrimerUsoCrearAdmin extends javax.swing.JPanel {
         return fechaObtenida;
     }
     
-    public int getIdentificacion(){
-        return Integer.parseInt(this.idTF.getText());
+    public long getIdentificacion(){
+        return Long.parseLong(this.idTF.getText());
     }
     
     public String getPassword(){
