+++ src/Vista/PanelPrimerUsoCrearPaciente.java
@@ -156,8 +156,8 @@ public class PanelPrimerUsoCrearPaciente extends javax.swing.JPanel {
         return fechaObtenida;
     }
     
-    public int getIdentificacion(){
-        return Integer.parseInt(this.idTF.getText());
+    public long getIdentificacion(){
+        return Long.parseLong(this.idTF.getText());
     }
     
     public String getPassword(){
